requirements.txt
streamlit
python-dotenv
langchain
langchain-google-genai
langchain-huggingface
langchain-chroma
chromadb
pypdf
sentence-transformers